+++ src/features/Keybinds.cpp
@@ -97,6 +97,13 @@ struct $modify(EditorUI) {
             fakeEditorPauseLayer(m_editorLayer)->saveLevel();
             Notification::create("Level saved", NotificationIcon::Success)->show();
         });
+        this->defineKeybind("pause-resume-playtest"_spr, [this]() {
+            // LevelEditorLayer::onPausePlaytest is inlined at least on Windows 
+            // but this does the job better probably anyway sooo
+            if (m_editorLayer->m_playbackMode != PlaybackMode::Not) {
+                this->onPlaytest(nullptr);
+            }
+        });
 
         this->defineKeybind("build-helper"_spr, [this]() {
             fakeEditorPauseLayer(m_editorLayer)->onBuildHelper(nullptr);
@@ -228,6 +235,14 @@ $execute {
         "",
         {},
         Category::EDITOR_MODIFY,
+        false
+    ));
+    BindManager::get()->registerBindable(BindableAction(
+        "pause-resume-playtest"_spr,
+        "Pause/Resume Playtest",
+        "Pauses or resumes the current playtest. Does not start playtesting",
+        {},
+        Category::EDITOR,
         false
     ));
     BindManager::get()->registerBindable(BindableAction(